Move test fixtures to top-level conftest.py
