+++ cubeviz/tests/test_ui.py
@@ -1,43 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 import os
 
-import pytest
+from .helpers import toggle_viewer, select_viewer
 
-from qtpy import QtCore
-from glue.app.qt import GlueApplication
-
-
-TEST_DATA_PATH = os.path.join(os.path.dirname(__file__), 'data')
-
-
-def toggle_viewer(qtbot, cubeviz_layout):
-    qtbot.mouseClick(
-        cubeviz_layout.button_toggle_image_mode, QtCore.Qt.LeftButton)
-
-def select_viewer(qtbot, viewer):
-    qtbot.mouseClick(viewer._widget, QtCore.Qt.LeftButton)
-
-@pytest.fixture(scope='module')
-def cubeviz_layout():
-    filename = os.path.join(TEST_DATA_PATH, 'data_cube.fits.gz')
-
-    app = GlueApplication()
-    app.run_startup_action('cubeviz')
-    app.load_data(filename)
-    app.setVisible(True)
-
-    return app.tab(0)
-
-@pytest.fixture(autouse=True)
-def reset_state(qtbot, cubeviz_layout):
-    # This yields the test itself
-    yield
-
-    # Make sure to return the application to this state between tests
-    if cubeviz_layout._single_viewer_mode:
-        toggle_viewer(qtbot, cubeviz_layout)
-    if cubeviz_layout._active_view is not cubeviz_layout.left_view:
-        select_viewer(qtbot, cubeviz_layout.left_view)
 
 def test_starting_state(cubeviz_layout):
     assert cubeviz_layout.isVisible() == True
